podcast-pipeline/requirements.txt
# Core
pyyaml>=6.0
python-dotenv>=1.0
requests>=2.31
httpx>=0.27
tenacity>=8.2          # retry logic
tqdm>=4.66             # progress bars

# RSS
feedparser>=6.0

# Audio download
yt-dlp>=2024.3         # broad podcast / media download support
aiohttp>=3.9           # async HTTP for concurrent downloads

# Transcription
assemblyai>=0.30       # AssemblyAI SDK
openai>=1.30           # OpenAI Whisper + DALL-E
deepgram-sdk>=3.5      # Deepgram SDK

# LLM
anthropic>=0.28        # Anthropic Claude
google-generativeai>=0.7

# Image generation
replicate>=0.27        # Replicate.com

# Telegram
python-telegram-bot>=21.0

# Storage / state
# (SQLite ships with Python's stdlib — no extra dep needed)

# Utilities
Pillow>=10.0           # image inspection / resizing
rich>=13.0             # beautiful terminal output
click>=8.1             # CLI interface
